requirements.txt
kubernetes
asyncssh
ptyprocess
escapism
traitlets
ruamel.yaml